+++ src/utils.test.js
@@ -15,6 +15,7 @@ import {
   switchAffiliation,
   setRecordForDeletion,
   parseEmptyFormValue,
+  sendCalloutOnAffiliationChange,
 } from './utils';
 import {
   CONTENT_TYPE_HEADER,
@@ -276,5 +277,89 @@ describe('parseEmptyFormValue', () => {
     const value = undefined;
 
     expect(parseEmptyFormValue(value)).toEqual(undefined);
+  });
+});
+
+describe('sendCalloutOnAffiliationChange', () => {
+  const callout = {
+    sendCallout: jest.fn(),
+  };
+  const tenantId = 'tenantId';
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  describe('When tenant changed', () => {
+    const okapi = {
+      tenant: 'okapiTenantId',
+    };
+
+    it('should send named informational notification callout', () => {
+      const stripes = {
+        user: {
+          user: {
+            tenants: [
+              {
+                id: tenantId,
+                name: 'name',
+              }
+            ],
+          },
+        },
+        okapi,
+      };
+      const expectedArgs = {
+        type: 'info',
+        message: (
+          <FormattedMessage
+            id="ui-inventory.affiliationChanging.namedNotification"
+            values={{ name: stripes.user.user.tenants[0].name }}
+          />
+        ),
+      };
+
+      sendCalloutOnAffiliationChange(stripes, tenantId, callout);
+
+      expect(callout.sendCallout).toHaveBeenCalledWith(expect.objectContaining(expectedArgs));
+    });
+
+    it('should send general informational notification callout', () => {
+      const stripes = {
+        user: {
+          user: {
+            tenants: [],
+          },
+        },
+        okapi,
+      };
+      const expectedArgs = {
+        type: 'info',
+        message: <FormattedMessage id="ui-inventory.affiliationChanging.notification" />,
+      };
+
+      sendCalloutOnAffiliationChange(stripes, tenantId, callout);
+
+      expect(callout.sendCallout).toHaveBeenCalledWith(expect.objectContaining(expectedArgs));
+    });
+  });
+
+  describe('When tenant is not changed', () => {
+    it('should not trigger callout', () => {
+      const stripes = {
+        user: {
+          user: {
+            tenants: [],
+          },
+        },
+        okapi: {
+          tenant: tenantId,
+        },
+      };
+
+      sendCalloutOnAffiliationChange(stripes, tenantId, callout);
+
+      expect(callout.sendCallout).not.toHaveBeenCalled();
+    });
   });
 });