+++ src/utils.js
@@ -930,3 +930,26 @@ export const redirectToMarcEditPage = (pathname, instance, location, history) =>
     search: searchParams.toString(),
   });
 };
+
+export const sendCalloutOnAffiliationChange = (stripes, tenantId, callout) => {
+  if (tenantId && stripes.okapi.tenant !== tenantId) {
+    const name = stripes.user.user.tenants.find(tenant => tenant.id === tenantId)?.name;
+
+    if (name) {
+      callout.sendCallout({
+        type: 'info',
+        message: (
+          <FormattedMessage
+            id="ui-inventory.affiliationChanging.namedNotification"
+            values={{ name }}
+          />
+        ),
+      });
+    } else {
+      callout.sendCallout({
+        type: 'info',
+        message: <FormattedMessage id="ui-inventory.affiliationChanging.notification" />,
+      });
+    }
+  }
+};